1.20.3 sources

+++ client/net/minecraft/world/entity/ai/goal/MoveToBlockGoal.java
@@ -64,9 +64,7 @@ public abstract class MoveToBlockGoal extends Goal {
    protected void moveMobToBlock() {
       this.mob
          .getNavigation()
-         .moveTo(
-            (double)((float)this.blockPos.getX()) + 0.5, (double)(this.blockPos.getY() + 1), (double)((float)this.blockPos.getZ()) + 0.5, this.speedModifier
-         );
+         .moveTo((double)this.blockPos.getX() + 0.5, (double)(this.blockPos.getY() + 1), (double)this.blockPos.getZ() + 0.5, this.speedModifier);
    }
 
    public double acceptedDistance() {
@@ -89,7 +87,7 @@ public abstract class MoveToBlockGoal extends Goal {
          this.reachedTarget = false;
          ++this.tryTicks;
          if (this.shouldRecalculatePath()) {
-            this.mob.getNavigation().moveTo((double)((float)var1.getX()) + 0.5, (double)var1.getY(), (double)((float)var1.getZ()) + 0.5, this.speedModifier);
+            this.mob.getNavigation().moveTo((double)var1.getX() + 0.5, (double)var1.getY(), (double)var1.getZ() + 0.5, this.speedModifier);
          }
       } else {
          this.reachedTarget = true;
